requirements.txt
Augmentor
autograd
cloudpickle
# cupy-cuda101==6.5.0
fastprogress
graphviz
image
imageio
imagesize
librosa
lmdb
matplotlib
matplotlib-venn
np-utils
numba
numexpr
numpy
# nvidia-ml-py3==7.352.0
opencv-contrib-python
opencv-python
openpyxl
pathlib
Pillow==8.2.0
pip-tools
pydot
pydot-ng
pydotplus
pygame
pyglet
Pygments
regex
requests==2.23.0
requests-oauthlib==1.3.0
scikit-image
scikit-learn
scipy
setuptools-git
sklearn==0.0
torch
torchsummary
torchtext
torchvision
tornado==4.5.3
tqdm==4.41.1
urllib3==1.24.3
